merge source

+++ server/api/v1/sys_auto_code.go
@@ -2,16 +2,19 @@ package v1
 
 import (
 	"fmt"
-	"gin-vue-admin/global/response"
+	"gin-vue-admin/global"
 	"gin-vue-admin/model"
+	"gin-vue-admin/model/response"
 	"gin-vue-admin/service"
 	"gin-vue-admin/utils"
 	"github.com/gin-gonic/gin"
+	"github.com/pkg/errors"
+	"go.uber.org/zap"
 	"net/url"
 	"os"
 )
 
-// @Tags SysApi
+// @Tags AutoCode
 // @Summary 自动代码模板
 // @Security ApiKeyAuth
 // @accept application/json
@@ -22,19 +25,12 @@ import (
 func CreateTemp(c *gin.Context) {
 	var a model.AutoCodeStruct
 	_ = c.ShouldBindJSON(&a)
-	AutoCodeVerify := utils.Rules{
-		"Abbreviation": {utils.NotEmpty()},
-		"StructName":   {utils.NotEmpty()},
-		"PackageName":  {utils.NotEmpty()},
-		"Fields":       {utils.NotEmpty()},
-	}
-	WKVerifyErr := utils.Verify(a, AutoCodeVerify)
-	if WKVerifyErr != nil {
-		response.FailWithMessage(WKVerifyErr.Error(), c)
+	if err := utils.Verify(a, utils.AutoCodeVerify); err != nil {
+		response.FailWithMessage(err.Error(), c)
 		return
 	}
 	if a.AutoCreateApiToSql {
-		apiList := [5]model.SysApi{
+		apiList := [6]model.SysApi{
 			{
 				Path:        "/" + a.Abbreviation + "/" + "create" + a.StructName,
 				Description: "新增" + a.Description,
@@ -44,6 +40,12 @@ func CreateTemp(c *gin.Context) {
 			{
 				Path:        "/" + a.Abbreviation + "/" + "delete" + a.StructName,
 				Description: "删除" + a.Description,
+				ApiGroup:    a.Abbreviation,
+				Method:      "DELETE",
+			},
+			{
+				Path:        "/" + a.Abbreviation + "/" + "delete" + a.StructName + "ByIds",
+				Description: "批量删除" + a.Description,
 				ApiGroup:    a.Abbreviation,
 				Method:      "DELETE",
 			},
@@ -67,23 +69,83 @@ func CreateTemp(c *gin.Context) {
 			},
 		}
 		for _, v := range apiList {
-			errC := service.CreateApi(v)
-			if errC != nil {
+			if err := service.AutoCreateApi(v); err != nil {
+				global.GVA_LOG.Error("自动化创建失败!请自行清空垃圾数据!", zap.Any("err", err))
 				c.Writer.Header().Add("success", "false")
-				c.Writer.Header().Add("msg", url.QueryEscape(fmt.Sprintf("自动化创建失败，%v，请自行清空垃圾数据", errC)))
+				c.Writer.Header().Add("msg", url.QueryEscape("自动化创建失败!请自行清空垃圾数据!"))
 				return
 			}
 		}
 	}
 	err := service.CreateTemp(a)
 	if err != nil {
-		response.FailWithMessage(fmt.Sprintf("创建失败，%v", err), c)
-		os.Remove("./ginvueadmin.zip")
+		if errors.Is(err, model.AutoMoveErr) {
+			c.Writer.Header().Add("success", "false")
+			c.Writer.Header().Add("msgtype", "success")
+			c.Writer.Header().Add("msg", url.QueryEscape(err.Error()))
+		} else {
+			c.Writer.Header().Add("success", "false")
+			c.Writer.Header().Add("msg", url.QueryEscape(err.Error()))
+			_ = os.Remove("./ginvueadmin.zip")
+		}
+
 	} else {
 		c.Writer.Header().Add("Content-Disposition", fmt.Sprintf("attachment; filename=%s", "ginvueadmin.zip")) // fmt.Sprintf("attachment; filename=%s", filename)对下载的文件重命名
 		c.Writer.Header().Add("Content-Type", "application/json")
 		c.Writer.Header().Add("success", "true")
 		c.File("./ginvueadmin.zip")
-		os.Remove("./ginvueadmin.zip")
+		_ = os.Remove("./ginvueadmin.zip")
+	}
+}
+
+// @Tags AutoCode
+// @Summary 获取当前数据库所有表
+// @Security ApiKeyAuth
+// @accept application/json
+// @Produce application/json
+// @Success 200 {string} string "{"success":true,"data":{},"msg":"获取成功"}"
+// @Router /autoCode/getTables [get]
+func GetTables(c *gin.Context) {
+	dbName := c.DefaultQuery("dbName", global.GVA_CONFIG.Mysql.Dbname)
+	err, tables := service.GetTables(dbName)
+	if err != nil {
+		global.GVA_LOG.Error("查询table失败!", zap.Any("err", err))
+		response.FailWithMessage("查询table失败", c)
+	} else {
+		response.OkWithDetailed(gin.H{"tables": tables}, "获取成功", c)
+	}
+}
+
+// @Tags AutoCode
+// @Summary 获取当前所有数据库
+// @Security ApiKeyAuth
+// @accept application/json
+// @Produce application/json
+// @Success 200 {string} string "{"success":true,"data":{},"msg":"获取成功"}"
+// @Router /autoCode/getDatabase [get]
+func GetDB(c *gin.Context) {
+	if err, dbs := service.GetDB(); err != nil {
+		global.GVA_LOG.Error("获取失败!", zap.Any("err", err))
+		response.FailWithMessage("获取失败", c)
+	} else {
+		response.OkWithDetailed(gin.H{"dbs": dbs}, "获取成功", c)
+	}
+}
+
+// @Tags AutoCode
+// @Summary 获取当前表所有字段
+// @Security ApiKeyAuth
+// @accept application/json
+// @Produce application/json
+// @Success 200 {string} string "{"success":true,"data":{},"msg":"获取成功"}"
+// @Router /autoCode/getColumn [get]
+func GetColumn(c *gin.Context) {
+	dbName := c.DefaultQuery("dbName", global.GVA_CONFIG.Mysql.Dbname)
+	tableName := c.Query("tableName")
+	if err, columns := service.GetColumn(tableName, dbName); err != nil {
+		global.GVA_LOG.Error("获取失败!", zap.Any("err", err))
+		response.FailWithMessage("获取失败", c)
+	} else {
+		response.OkWithDetailed(gin.H{"columns": columns}, "获取成功", c)
 	}
 }
